requirements.txt
flask==2.3.3
flask-cors==4.0.0
instagrapi==1.19.3
python-dotenv==1.0.0